requirements.txt
python-docx>=0.8.11
click>=8.1.3
python-dotenv>=1.0.0
ollama>=0.1.72
pypdf>=3.0.0
pandas>=1.3.0       # Для чтения Excel
openpyxl>=3.0.0     # Движок для pandas для работы с .xlsx